Add and handle exception for malformed utt2spk.
Some stages in kaldi (in particular a cleanup stage of the librispeech
recipe) add datasets with modified utterance ids in utt2spk which do not
correspond to the wav.scp files. Since we do not intend to support this,
an exception is raised. However, since this might occur, the treatment
is a skip of the affected datasets.