Add catch for improper egs path in KaldiDatabase
Without this assert, an improper egs path (e.g. due to empty $KALDI_ROOT
env variable) results in an empty database dict.
This usually breaks down when a lazy dataset on the database is
instantiated - because of trying to concatenate an empty list of datasets -
which is very late and obscure.
The check results in an early noisy breakage.